Brazilian Inflation and Core Inflation Data

A dataset containing the variation of the Brazilian headline Consumer Price Index (CPI) along with core inflation measures.

coreinf_br

Format

An object of class (inherits from tbl, data.frame) with 302 rows and 9 columns. Monthly % variation from 1994-07-01 to 2019-08-01.

ipca

Broad National Consumer Price Index. ipca represents the variation in consumer price index. It is the reference for the inflation targeting system implemented in June 1999.

ipcams

Broad national consumer price index - Core IPCA trimmed means smoothed. Excludes items whose monthly change is higher than the 80th percentile or bellow the 20th percentile with smoothing of items with low frequency variations.

ipcama

Broad national consumer price index - Core IPCA trimmed means non smoothed. The same as ipcams but without smoothing.

ipcaex0

Broad national consumer price index - Core exclusion - EX0. Excludes food at home and administered goods from ipca inflation.

ipcaex1

Broad national consumer price index - Core exclusion - EX1. Excludes 10 of the 16 items in the food at home group as well as domestic fuels and vehicle fuels.

ipcadp

Broad national consumer price index - Core IPCA - Double Weight. Adjusts original weigths of each item according to its relative volatility, which reduces the weight of components with more volatile inflation.

ipcaex2

Broad national consumer price index - Core exclusion - EX2. Considers only the underlying items of the services, industrial goods, and food away from home groups.

ipcaex3

Broad national consumer price index - Core exclusion - EX3. Considers only the underlying items of the services and industrial goods groups.

Source

Series can be obtained from function getsgs or by codes: 433, 11427, 16121, 27838, 27839, 11426, 4466, 16122 in the
Time Series Management System

References

Central Bank of Brazil (2018). New core inflation measures. Inflation Report, june 2018.
Inflation report boxes: New core inflation measures / 2018-06

Tito N. T da Silva Filho and Francisco M. R. Figueiredo (2014). Revisiting the Core Inflation Measures of the Banco Central do Brasil. Working Papers Series 356, Central Bank of Brazil, Research Department.
https://www.bcb.gov.br/pec/wps/port/TD356.pdf