Map over Several Horizons to Extract Dynamic Adjustment Parameter

dyn_adj is a fast way to extract the parameter of the dynamic adjustment for a range of horizons.

dyn_adj(inf, core, H, p, ...)

Arguments

inf
A vector containing headline inflation.

core
A vector containing a measure of core inflation.

H
An integer that gives the maximum horizon.

p
An integer, lag in the regression (see details).

...
Additional parameters.

Value

A tibble with the dynamic adjustment parameter and p-value for the t test of \(\lambda^j_h = 0\) for \(j = ( , *)\). Row 1 of the tibble is \(h = 1\), row 2 is \(h = 2\) and so on. For estimating \(\lambda^*_h\), put core as the first argument of the function and headline inflation as the second.

Details

For a model \(\pi_{t + h} - \pi_t = a_0 + \lambda_h(\pi_t - \pi^*_t) + \sum_{i = 1}^pa_i\pi_{t - i} + e_{t + h}\), where \(\pi\) is the headline inflation and \(\pi^*\) a measure of core inflation, this function gives \(\lambda_h\) for \(h \in 1, 2 ..., H\).

In core inflation's literature, parameters \(\lambda_h\) and \(\lambda^*_h\) of the following regressions are important:

\(\pi_{t + h} - \pi_t = a_0 + \lambda_h(\pi_t - \pi^*_t) + \sum_{i = 1}^pa_i\pi_{t - i} + e_{t + h}\),

\(\pi^*_{t + h} - \pi^*_t = a^*_0 + \lambda^*_h(\pi_t - \pi^*_t) + \sum_{i = 1}^pa^*_i\pi^*_{t - i} + e^*_{t + h}\).

A good core inflation measure should imply \(\lambda_h < 0\) and \(\lambda^*_h = 0\).

See also

Examples

inf_head <- coreinf_br[["ipca"]]
inf_corems <- coreinf_br[["ipcams"]]
inf_coredp <- coreinf_br[["ipcadp"]]
dyn_adj(inf_head, inf_corems, 4, 5, recursive = TRUE)

#> # A tibble: 4 x 2
#>    alpha    val_p
#>    <dbl>    <dbl>
#> 1 -0.534 2.38e-15
#> 2 -0.805 4.35e-19
#> 3 -0.834 6.10e-19
#> 4 -0.918 1.58e-21

out <- purrr::map(list(inf_corems, inf_coredp), ~ dyn_adj(inf_head, .x, 4, 5, recursive = TRUE))
purrr::map(out, dplyr::summarise_all, mean)

#> [[1]]
#> # A tibble: 1 x 2
#>    alpha    val_p
#>    <dbl>    <dbl>
#> 1 -0.773 5.96e-16
#> 
#> [[2]]
#> # A tibble: 1 x 2
#>    alpha        val_p
#>    <dbl>        <dbl>
#> 1 -0.938 0.0000000998
#>