Select Best Model

A function for selecting the best model fitted by the function model_est or model_est_single.

model_best(models, ic = BIC)

Arguments

models
An object generated by the funcion model_est or model_est_single.

ic
Information criterion: BIC or AIC.

Value

A list containing the following elements:

data

Tibble with the data used in fitting the model.

model

An lm object.

n_obs

Number of observations of the dependent variable.

h

Horizon used in direct estimation.

See also

Examples

inf_head <- coreinf_br[["ipca"]]
inf_corems <- coreinf_br[["ipcams"]]
pq <- lags(2, 1)
model1 <- model_est(inf_head, inf_corems, 2, pq)
model_best(model1)

#> $data
#> # A tibble: 302 x 3
#>    inf_head inf_head_lag2 inf_corems_lag2
#>       <dbl>         <dbl>           <dbl>
#>  1     6.84         NA                 NA
#>  2     1.86         NA                 NA
#>  3     1.53          6.84              NA
#>  4     2.62          1.86              NA
#>  5     2.81          1.53              NA
#>  6     1.71          2.62              NA
#>  7     1.7           2.81              NA
#>  8     1.02          1.71              NA
#>  9     1.55          1.7               NA
#> 10     2.43          1.02              NA
#> # ... with 292 more rows
#> 
#> $model
#> 
#> Call:
#> stats::lm(formula = inf_head ~ inf_head_lag2 + inf_corems_lag2)
#> 
#> Coefficients:
#>     (Intercept)    inf_head_lag2  inf_corems_lag2  
#>          0.2223           0.2169           0.3430  
#> 
#> 
#> $n_obs
#> [1] 302
#> 
#> $h
#> [1] 2
#> 

p <- unique(lags(2, 1)[, 1:2])
model2 <- model_est_single(inf_head, 2, p)
model_best(model2, ic = AIC)

#> $data
#> # A tibble: 302 x 3
#>    inf_head inf_head_lag2 inf_head_lag3
#>       <dbl>         <dbl>         <dbl>
#>  1     6.84         NA            NA   
#>  2     1.86         NA            NA   
#>  3     1.53          6.84         NA   
#>  4     2.62          1.86          6.84
#>  5     2.81          1.53          1.86
#>  6     1.71          2.62          1.53
#>  7     1.7           2.81          2.62
#>  8     1.02          1.71          2.81
#>  9     1.55          1.7           1.71
#> 10     2.43          1.02          1.7 
#> # ... with 292 more rows
#> 
#> $model
#> 
#> Call:
#> stats::lm(formula = inf_head ~ inf_head_lag2 + inf_head_lag3)
#> 
#> Coefficients:
#>   (Intercept)  inf_head_lag2  inf_head_lag3  
#>        0.2330         0.3810         0.1972  
#> 
#> 
#> $n_obs
#> [1] 302
#> 
#> $h
#> [1] 2
#>