Error Prediction Two Variables

This function calculates the root mean square error or mean absolute error of models estimated by the function model_fcast. This function is appropriate for forecasting evaluation of a core inflation measure that is never revised.

model_error(k, y, xreg, lags, h, RMSE = TRUE)

Arguments

k
An integer to determine the number of observations to include in the out-of-sample set.

y
A numeric vector or time series of the dependet variable; tipplically the headline inflation.

xreg
A numeric vector or time series of the independent variable; a measure of core inflation that is never revised.

lags
A data frame generated by the function lags.

h
An integer to set the desired horizon.

RMSE
Logical: if TRUE returns RMSE, if FALSE returns MAE.

Value

An integer contaning the calculated error.

See also

Examples

inf_head <- coreinf_br[["ipca"]]
inf_corems <- coreinf_br[["ipcams"]]
pq <- lags(2, 1)
model_error(24, inf_head, inf_corems, pq, 2, RMSE = TRUE)

#> [1] 0.3364706

purrr::map_dbl(1:12, ~ model_error(24, inf_head, inf_corems, pq, .x, RMSE = FALSE))

#>  [1] 0.2231067 0.2388609 0.2409641 0.2483008 0.2601510 0.2783094 0.2837708
#>  [8] 0.2655851 0.2520037 0.2584306 0.2706968 0.2696325