Estimate Direct Forecasting Model

Given y and xreg, this function estimates a direct forecasting model as described below, in section "Details".

model_est(y, xreg, h, lags)

Arguments

y
A numeric vector or time series.

xreg
A numeric vector for the exogenous variable.

h
An integer. The horizon of prediction.

lags
A data frame generated by the function lags.

Value

A list with following components:

model

A lm object.

n_obs

Number of observations of y.

h

The horizon from which y is being forecasted.

Details

Condiser the following regression:

\(y_{i,t+h} = a_0 + y_{i,t + 1 - pstart} + ... + y_{i,t + 1 - pend} + x_{i,t + 1 - qstart} + ... + x_{i,t + 1 - qend} + e_{i,t + h}\).

model_est estimates the above regression for all indexes in the data frame lags (p_start \(\ge 1\) and q_start \(\ge 1\)). Forecasting from a model with this structure is called direct forecast, where \(h\) denotes the horizon over which \(y\) is to be forecast. This approach can be used to forecast several steps ahead.

A measure of core inflation (\(\pi^*\)) should improve forecasting of headline inflation (\(\pi\)). A way of test this property is running a regression like that one above:

\(\pi_{t+h} = \alpha_0 + \sum_{i = 1}^p\beta_i\pi_{t-i+1} + \sum_{i = 1}^q\gamma_i\pi^*_{t-i+1} e_{t + h}\).

Using recursive = FALSE (default) in function lags is equivalent to estimate the previous equation for headline inflation. If recursive = TRUE, \(i\) is not necessarily initiated in 1, i.e, p_start and q_start also changes in the first equation. Allowing p_start and q_start to vary, there are more models to be estimated because it opens up a lot of new possibilities.

See also

Examples

inf_head <- coreinf_br[["ipca"]]
inf_corems <- coreinf_br[["ipcams"]]
pq <- lags(2, 1)
model_est(inf_head, inf_corems, 2, pq)

#> [[1]]
#> [[1]]$data
#> # A tibble: 302 x 3
#>    inf_head inf_head_lag2 inf_corems_lag2
#>       <dbl>         <dbl>           <dbl>
#>  1     6.84         NA                 NA
#>  2     1.86         NA                 NA
#>  3     1.53          6.84              NA
#>  4     2.62          1.86              NA
#>  5     2.81          1.53              NA
#>  6     1.71          2.62              NA
#>  7     1.7           2.81              NA
#>  8     1.02          1.71              NA
#>  9     1.55          1.7               NA
#> 10     2.43          1.02              NA
#> # ... with 292 more rows
#> 
#> [[1]]$model
#> 
#> Call:
#> stats::lm(formula = inf_head ~ inf_head_lag2 + inf_corems_lag2)
#> 
#> Coefficients:
#>     (Intercept)    inf_head_lag2  inf_corems_lag2  
#>          0.2223           0.2169           0.3430  
#> 
#> 
#> [[1]]$n_obs
#> [1] 302
#> 
#> [[1]]$h
#> [1] 2
#> 
#> 
#> [[2]]
#> [[2]]$data
#> # A tibble: 302 x 4
#>    inf_head inf_head_lag2 inf_head_lag3 inf_corems_lag2
#>       <dbl>         <dbl>         <dbl>           <dbl>
#>  1     6.84         NA            NA                 NA
#>  2     1.86         NA            NA                 NA
#>  3     1.53          6.84         NA                 NA
#>  4     2.62          1.86          6.84              NA
#>  5     2.81          1.53          1.86              NA
#>  6     1.71          2.62          1.53              NA
#>  7     1.7           2.81          2.62              NA
#>  8     1.02          1.71          2.81              NA
#>  9     1.55          1.7           1.71              NA
#> 10     2.43          1.02          1.7               NA
#> # ... with 292 more rows
#> 
#> [[2]]$model
#> 
#> Call:
#> stats::lm(formula = inf_head ~ inf_head_lag2 + inf_head_lag3 + 
#>     inf_corems_lag2)
#> 
#> Coefficients:
#>     (Intercept)    inf_head_lag2    inf_head_lag3  inf_corems_lag2  
#>        0.222407         0.216054         0.002489         0.341054  
#> 
#> 
#> [[2]]$n_obs
#> [1] 302
#> 
#> [[2]]$h
#> [1] 2
#> 
#>