Direct Autoregressive Forecast Model

Estimates a direct autoregressive forecast model.

model_est_single(y, h, lags)

Arguments

y
A vector or a time series.

h
An integer, the horizon from which y is being forecasted.

lags
A data frame with lags to begin and to stop. These lags can be indirectly obtaneid by the function lags as example below shows.

Value

A list of the following components:

model

A lm object.

n_obs

Number of observations.

h

The horizon.

Details

Condiser the following regression:

\(y_{i,t+h} = a_0 + y_{i,t + 1 - pstart} + ... + y_{i,t + 1 - pend} + e_{i,t + h}\).

model_est_single estimates the above regression for all indexes in the data frame lags (p_start \(\ge 1\)).

Details section for the function model_est provides more information. Forecasting obtainded by the direct autoregressive model can be compared with forecasting where lags of the core inflation measure are considered in the model as in model_est function.

See also

Examples

x <- arima.sim(n = 30, list(ar = c(0.8897, -0.4858)))
p <- unique(lags(2, 1)[, 1:2])
model_est_single(x, 2, p)

#> [[1]]
#> [[1]]$data
#> # A tibble: 30 x 2
#>         x x_lag2
#>     <dbl>  <dbl>
#>  1  0.321 NA    
#>  2  0.381 NA    
#>  3  0.546  0.321
#>  4 -1.00   0.381
#>  5 -0.421  0.546
#>  6  2.00  -1.00 
#>  7  1.89  -0.421
#>  8 -0.229  2.00 
#>  9 -1.14   1.89 
#> 10 -1.73  -0.229
#> # ... with 20 more rows
#> 
#> [[1]]$model
#> 
#> Call:
#> stats::lm(formula = x ~ x_lag2)
#> 
#> Coefficients:
#> (Intercept)       x_lag2  
#>      0.1384      -0.1995  
#> 
#> 
#> [[1]]$n_obs
#> [1] 30
#> 
#> [[1]]$h
#> [1] 2
#> 
#> 
#> [[2]]
#> [[2]]$data
#> # A tibble: 30 x 3
#>         x x_lag2 x_lag3
#>     <dbl>  <dbl>  <dbl>
#>  1  0.321 NA     NA    
#>  2  0.381 NA     NA    
#>  3  0.546  0.321 NA    
#>  4 -1.00   0.381  0.321
#>  5 -0.421  0.546  0.381
#>  6  2.00  -1.00   0.546
#>  7  1.89  -0.421 -1.00 
#>  8 -0.229  2.00  -0.421
#>  9 -1.14   1.89   2.00 
#> 10 -1.73  -0.229  1.89 
#> # ... with 20 more rows
#> 
#> [[2]]$model
#> 
#> Call:
#> stats::lm(formula = x ~ x_lag2 + x_lag3)
#> 
#> Coefficients:
#> (Intercept)       x_lag2       x_lag3  
#>      0.1339       0.2808      -0.8340  
#> 
#> 
#> [[2]]$n_obs
#> [1] 30
#> 
#> [[2]]$h
#> [1] 2
#> 
#>