Map RMSE and MAE by Combining Two Data Frames

rmse_mae is a functional to compute root mean square error and mean absolute error for all variables in a data frame with respect to all variables in another data frame. The number of rows of both data frames must be equal.

rmse_mae(dfa, dfb)

Arguments

dfa
A data frame of numeric vectors.

dfb
A data frame of numeric vectors. The number of rows of dfb must be equal to dfa.

Value

A data frame

Examples

# Measuring how far wavelet core inflation measures are
# from moving averages of headline inflation.

inf_head <- coreinf_br[["ipca"]]

mavc_order <- c(13, 25, 37)
names(mavc_order) <- paste0("mavc", mavc_order)
df_mavc <- purrr::map_df(mavc_order, ~mavc(x = inf_head, k = .))

corewav1 <- wmtsa::wavShrink(inf_head)
corewav2 <- wmtsa::wavShrink(inf_head, wavelet = "haar")
df_core <- tibble::tibble(corewav1, corewav2)

rmse_mae(df_core, df_mavc)

#>                           rmse       mae
#> corewav1_mavc13_rmse 0.3345918 0.1594345
#> corewav2_mavc13_rmse 0.3216404 0.1360057
#> corewav1_mavc25_rmse 0.3574015 0.1552253
#> corewav2_mavc25_rmse 0.3465084 0.1367323
#> corewav1_mavc37_rmse 0.3821001 0.1712084
#> corewav2_mavc37_rmse 0.3726842 0.1554074