Wavelet Denoising

The basic idea behind smooth_wavelet is to denoise a vector of data using wavelet methods that are available in 3 R packages: wmtsa, EbayesThresh and wavethresh. Do not mix parameters related to different methods.

smooth_wavelet(x, thfun = wavShrink, wtfun = NULL,
  wtfunlist = list(), ...)

Arguments

x
A numeric vector. If the data is not of length \(2 ^ J\) for some integer \(J\), the function WiSEBoot::padVector increases the length of data to achieve the particular length requirement based on reflection type. See the help page of the mentioned function for details. After the signal computation, the length is again adjusted for the orginal length of x.

thfun
A function to denoise x. This should be a method available in packages wmtsa, EbayesThresh and wavethresh. The options are wavShrink,
ebayesthresh.wavelet and threshold. Default: wavShrink.

wtfun
A wavelet transform function: dwt or modwt if thfun = ebayesthresh.wavelet. wd if thfun = threshold. If thfun = wavShrink, wtfun = NULL (default).

wtfunlist
A named list of parameters to pass to functions dwt, modwt or wd.

...
Additional arguments to pass to functions:

wmtsa::wavShrink

EbayesThresh::ebayesthresh.wavelet

wavethresh::threshold

Value

A numeric vector with the wavelet-based signal estimation.

See also

Examples

inf_head <- coreinf_br[["ipca"]]

# From package wmtsa (default):
smooth_wavelet(inf_head)

#>   [1]  6.82272324  1.78733835  1.70976638  2.61721035  2.49844330  2.04360559
#>   [7]  1.39824760  1.21994522  1.67336192  2.31819330  2.54617064  2.32272093
#>  [13]  1.88705424  1.34734724  1.08527759  1.32781450  1.58811427  1.54649544
#>  [19]  1.31653594  0.96168833  0.77173429  0.98554784  1.18973703  1.06611613
#>  [25]  0.85720192  0.63185947  0.40070419  0.39774325  0.51970258  0.57564628
#>  [31]  0.62644328  0.63696143  0.58822485  0.52382578  0.42439574  0.33498415
#>  [37]  0.26045932  0.19220494  0.18305977  0.22091533  0.29593056  0.38869828
#>  [43]  0.44691981  0.46267479  0.43164919  0.35622912  0.23976593  0.08722175
#>  [49] -0.05505451 -0.14761929 -0.14314040 -0.01050261  0.16508656  0.34672313
#>  [55]  0.61780332  0.87424434  0.86557438  0.63513046  0.40760224  0.32342512
#>  [61]  0.91318943  0.58722809  0.66509350  0.73353553  0.69683755  0.60290357
#>  [67]  0.50092228  0.36703353  0.30996683  0.31342234  0.23247942  0.34627124
#>  [73]  1.42625280  0.96271098  0.43702981  0.19225241  0.40602438  0.56531707
#>  [79]  0.56389081  0.55371499  0.57031609  0.60688841  0.58055630  0.44358791
#>  [85]  0.87486769  0.52695395  0.58668383  0.61862003  0.62840906  0.63402360
#>  [91]  0.63991991  0.62411885  0.58149526  0.55657473  0.52279668  0.47282617
#>  [97]  1.01254120  0.63038165  0.72706442  1.24803487  2.74563101  2.31432661
#> [103]  1.99643901  1.58517425  1.22970213  0.86149543  0.46848797  0.20531458
#> [109]  0.24745177  0.41003968  0.47387047  0.49177517  0.50562367  0.50288803
#> [115]  0.51228236  0.53379093  0.55958788  0.58400288  0.59499651  0.59313538
#> [121]  0.58479828  0.57227812  0.55463470  0.53899467  0.53218327  0.52774446
#> [127]  0.53034886  0.53196914  0.51207799  0.48602021  0.45280757  0.41959535
#> [133]  0.41313074  0.41475459  0.41979054  0.42491785  0.41191398  0.39269179
#> [139]  0.37297803  0.35292659  0.33827180  0.32689440  0.31749056  0.31093919
#> [145]  0.30753634  0.30672537  0.30759901  0.30969723  0.31278689  0.31662646
#> [151]  0.32129241  0.32650864  0.33175001  0.33729778  0.34323622  0.34967987
#> [157]  0.35672935  0.36376956  0.37034369  0.37629595  0.38145517  0.38591291
#> [163]  0.39001875  0.39382167  0.39732599  0.40077894  0.40417218  0.40726305
#> [169]  0.41015109  0.41244983  0.41403177  0.41664841  0.42067473  0.42511795
#> [175]  0.42966861  0.43214043  0.43038989  0.42863897  0.42882652  0.42260024
#> [181]  0.41196395  0.39562964  0.37405741  0.38249970  0.42994450  0.50533643
#> [187]  0.59350511  0.63389853  0.59505754  0.49015311  0.35242759  0.24041425
#> [193]  0.20260634  0.25076258  0.35314903  0.47913848  0.58711636  0.63484429
#> [199]  0.63306287  0.58460457  0.50309247  0.43608405  0.38937486  0.37010165
#> [205]  0.38969623  0.41768269  0.44611993  0.47158270  0.47920500  0.48147916
#> [211]  0.48263541  0.48078964  0.48047075  0.47904861  0.47612654  0.47427789
#> [217]  0.47507373  0.47643872  0.47865740  0.48716140  0.50341780  0.53126303
#> [223]  0.56367250  0.57792969  0.56309504  0.51492500  0.44404567  0.38578268
#> [229]  0.36396543  0.38439298  0.44688417  0.52665029  0.59834342  0.65787440
#> [235]  0.69206807  0.69388688  0.66484331  0.59661428  0.49893325  0.40331338
#> [241]  0.33560652  0.32748912  0.39583398  0.52301690  0.69369668  0.86246238
#> [247]  0.98581337  1.04356823  1.00835009  0.90072271  0.75523923  0.60980591
#> [253]  0.52045273  0.51391364  0.58928239  0.71446280  0.83063372  0.88658635
#> [259]  0.86204790  0.77508544  0.66856804  0.57854205  0.52013508  0.48650368
#> [265]  0.46059237  0.43381335  0.40498846  0.37851824  0.35706565  0.33924221
#> [271]  0.32454160  0.31296096  0.30375782  0.29736854  0.29305268  0.28830325
#> [277]  0.28477739  0.28418829  0.28433440  0.28557270  0.29585380  0.30471161
#> [283]  0.30574583  0.31256473  0.29786731  0.19849753  0.12621739  0.75376812
#> [289]  0.27516579  0.22550528  0.26890711  0.30883942  0.30431124  0.30888533
#> [295]  0.30949182  0.30686269  0.30169303  0.29606535  0.29415346  0.29503284
#> [301]  0.29720083  0.29754699
#> attr(,"wavelet")
#> [1] "s8"
#> attr(,"n.level")
#> [1] 8
#> attr(,"shrink.fun")
#> [1] "hard"
#> attr(,"thresh.fun")
#> [1] "universal"
#> attr(,"thresh.scale")
#> [1] 1
#> attr(,"transform")
#> [1] "modwt"
#> attr(,"noise.variance")
#> [1] -1
#> attr(,"reflect")
#> [1] TRUE

# From package EbayesThresh:
smooth_wavelet(inf_head, ebayesthresh.wavelet, modwt, list(wf = "haar"),
               a = NA, vscale = "independent")

#>   [1] 6.68402295 1.61245955 1.46610409 2.38710942 2.40783461 1.52697313
#>   [7] 1.56970225 1.48151022 1.49684089 2.17293326 2.08091998 2.05422978
#>  [13] 2.20689588 0.97914476 1.27109241 1.34954126 1.29142716 1.29499235
#>  [19] 1.22925016 1.16061735 0.94497461 1.24893505 1.03915168 1.04903006
#>  [25] 0.98348319 0.69143869 0.63167306 0.65494988 0.61549675 0.62535156
#>  [31] 0.59507812 0.56140625 0.52941406 0.49871094 0.46523437 0.43292969
#>  [37] 0.40050781 0.36726562 0.34039062 0.32054687 0.30765625 0.30207031
#>  [43] 0.29703125 0.28820312 0.27769531 0.27003906 0.26453125 0.26093750
#>  [49] 0.26378906 0.22075383 0.20469599 0.22969599 0.24373647 0.37890850
#>  [55] 0.50548228 0.52938853 0.53346807 0.49680003 0.47976562 0.35845386
#>  [61] 0.66076489 0.53437500 0.43372825 0.67697487 0.57066406 0.57199219
#>  [67] 0.56890625 0.56171875 0.55562500 0.55253906 0.39637197 0.07120059
#>  [73] 1.18454997 1.10308909 0.16667176 0.40768694 0.55167969 0.55417969
#>  [79] 0.55738281 0.55816406 0.56164062 0.57050781 0.57972656 0.59097656
#>  [85] 0.60378906 0.61886719 0.63832031 0.64522880 0.64350826 0.64946403
#>  [91] 0.65767847 0.66663012 0.69901293 0.71160682 0.67384555 0.67098675
#>  [97] 0.71526573 0.68689133 0.75132119 1.01445787 2.62559167 1.80733582
#> [103] 1.83283382 1.55229882 1.15524588 0.89351619 0.51461494 0.38449546
#> [109] 0.46582607 0.51154609 0.55319339 0.55701753 0.55530057 0.56197124
#> [115] 0.57361769 0.58301728 0.58265625 0.56699219 0.55730469 0.55378906
#> [121] 0.55378906 0.55363281 0.55328125 0.55503906 0.55437500 0.55078125
#> [127] 0.54261719 0.53335937 0.51980469 0.50355469 0.48273437 0.46187500
#> [133] 0.44484375 0.42890625 0.41496094 0.40199219 0.38765625 0.37402344
#> [139] 0.36125000 0.34625000 0.33085937 0.31574219 0.30417969 0.29480469
#> [145] 0.29054687 0.28867187 0.29269531 0.29910156 0.30476562 0.31085937
#> [151] 0.31601562 0.32218750 0.33074219 0.34062500 0.35105469 0.36261719
#> [157] 0.37535156 0.38929687 0.40148437 0.41453125 0.42656250 0.43714844
#> [163] 0.44402344 0.44933594 0.45351562 0.45609375 0.45683594 0.45406250
#> [169] 0.44832031 0.44148437 0.43488281 0.43027344 0.42628906 0.42246094
#> [175] 0.42050781 0.41718750 0.41074219 0.40425781 0.39707031 0.39007812
#> [181] 0.38511719 0.38339844 0.38511719 0.39128906 0.40128906 0.41324219
#> [187] 0.42640625 0.43656250 0.44296875 0.44734375 0.44980469 0.45285156
#> [193] 0.45972656 0.46933594 0.48246094 0.49609375 0.50703125 0.51324219
#> [199] 0.51761719 0.51785156 0.51324219 0.50574219 0.49527344 0.48628906
#> [205] 0.48148437 0.47980469 0.48054687 0.48218750 0.48281250 0.48125000
#> [211] 0.47808594 0.47375000 0.46984375 0.46765625 0.46386719 0.46187500
#> [217] 0.46460937 0.46917969 0.47714844 0.48464844 0.49082031 0.49652344
#> [223] 0.50011719 0.50085937 0.50015625 0.49785156 0.49417969 0.49203125
#> [229] 0.49125000 0.49304687 0.49851562 0.50750000 0.51710937 0.52933594
#> [235] 0.53875000 0.54898437 0.55921875 0.56703125 0.57355469 0.58195312
#> [241] 0.59277344 0.60929687 0.62976562 0.65218750 0.67585937 0.69734375
#> [247] 0.71605469 0.72949219 0.73699219 0.73832031 0.73941406 0.73718750
#> [253] 0.73250000 0.72726562 0.72410156 0.72000000 0.71234375 0.69781250
#> [259] 0.67730469 0.65031250 0.61703125 0.58332031 0.54863281 0.51332031
#> [265] 0.48167969 0.45222656 0.42382812 0.39882812 0.37613281 0.35480469
#> [271] 0.33285156 0.31160156 0.29589844 0.28347656 0.27335937 0.26765625
#> [277] 0.26902344 0.26937500 0.27121094 0.27609375 0.28074219 0.28816406
#> [283] 0.29609375 0.30257812 0.30761719 0.31339844 0.25700870 0.56783212
#> [289] 0.16920840 0.34149890 0.35090077 0.35687347 0.35675235 0.36151404
#> [295] 0.36611948 0.39016256 0.41829877 0.42560097 0.40394993 0.34180484
#> [301] 0.26931953 0.14022537

# From package wavethresh:
smooth_wavelet(inf_head, threshold, wd, list(filter.number = 8), policy = "cv")

#>   [1]  5.87363727  1.65461880  1.52698432  2.19138736  2.48671295  1.85141804
#>   [7]  1.57267004  1.35400802  1.55991500  1.95547008  2.26861054  2.13675809
#>  [13]  1.79186550  1.21553637  1.18472848  1.34722691  1.63258264  1.54248342
#>  [19]  1.27094430  1.00659372  0.82136370  1.04334661  1.02385927  0.98091835
#>  [25]  0.78432139  0.55962141  0.37894474  0.34530909  0.40502032  0.48797181
#>  [31]  0.56091178  0.50645690  0.45874847  0.50460499  0.40298378  0.36418110
#>  [37]  0.31309480  0.30462148  0.32261880  0.34569311  0.38351207  0.44239103
#>  [43]  0.50313544  0.54522291  0.53637737  0.46260910  0.33331588  0.14501768
#>  [49] -0.01070743 -0.01272362  0.08364492  0.18909686  0.31640236  0.43849899
#>  [55]  0.55063611  0.75663810  0.89242058  0.64608078  0.32504700  0.44178192
#>  [61]  0.74589827  0.72854962  0.55539514  0.83846877  0.65107284  0.62673355
#>  [67]  0.49168883  0.40047478  0.32271268  0.23638696  0.18419691  0.42545867
#>  [73]  1.09177978  0.88347200  0.46256067  0.44317899  0.60342984  0.57668603
#>  [79]  0.47748931  0.46182411  0.49660865  0.52156243  0.55584518  0.66399001
#>  [85]  0.76480588  0.69486843  0.57891040  0.69485752  0.75544472  0.77052694
#>  [91]  0.71425761  0.69371932  0.67473554  0.57503499  0.45796669  0.53067376
#>  [97]  0.69255462  0.55201386  0.48595960  1.29951012  2.29379675  1.91529832
#> [103]  1.83091807  1.37345076  1.04382822  0.77702033  0.60836326  0.31041217
#> [109]  0.27266050  0.34284311  0.50693105  0.54128214  0.50881073  0.52682929
#> [115]  0.54758590  0.50135074  0.45977784  0.49495581  0.57782391  0.69990310
#> [121]  0.78759440  0.71177079  0.58750584  0.60615400  0.67634696  0.64640159
#> [127]  0.57381300  0.54717114  0.52729776  0.46872891  0.39529677  0.33025465
#> [133]  0.30099402  0.33475869  0.40960221  0.49593758  0.56091330  0.56927795
#> [139]  0.53452910  0.48277954  0.42816401  0.38843651  0.36296567  0.34261595
#> [145]  0.33341778  0.33991278  0.35688768  0.38005947  0.39781987  0.39801227
#> [151]  0.38576097  0.36908607  0.35607557  0.35565589  0.36439091  0.37594214
#> [157]  0.38597781  0.38970715  0.38897453  0.38744632  0.38716207  0.39039460
#> [163]  0.39614711  0.40237198  0.40826026  0.41281801  0.41614057  0.41892869
#> [169]  0.42073820  0.42120053  0.42080523  0.41974758  0.41883179  0.41928227
#> [175]  0.41967425  0.41836519  0.41620655  0.41271735  0.40931214  0.40903288
#> [181]  0.40501058  0.39284799  0.37657387  0.35748733  0.35596645  0.38968308
#> [187]  0.44370668  0.50356539  0.53322278  0.49740532  0.42144224  0.33721058
#> [193]  0.28610908  0.31215953  0.39070860  0.48244606  0.56064255  0.59349473
#> [199]  0.58857630  0.56910334  0.53079480  0.47362291  0.41000046  0.34637628
#> [205]  0.31154306  0.33397651  0.39516054  0.47022577  0.53128069  0.54751841
#> [211]  0.52995984  0.50093772  0.47077086  0.45215137  0.44628265  0.44741161
#> [217]  0.45896896  0.48504322  0.51862346  0.55066211  0.58053219  0.60453172
#> [223]  0.61898457  0.62672434  0.60868353  0.54810586  0.46610134  0.38437054
#> [229]  0.33781774  0.36310922  0.43816036  0.52808860  0.61015654  0.65705050
#> [235]  0.67051467  0.66705256  0.63888956  0.57990000  0.50893630  0.44368573
#> [241]  0.40386715  0.41419976  0.46229516  0.52158138  0.60293154  0.70838605
#> [247]  0.81702392  0.92009143  0.95319926  0.85736034  0.68095588  0.47972284
#> [253]  0.34558232  0.37247429  0.51105661  0.68661165  0.82370969  0.84064787
#> [259]  0.76813839  0.66446067  0.56805195  0.52260256  0.51595525  0.51870756
#> [265]  0.51989324  0.50652489  0.48298437  0.46054557  0.44140303  0.42868056
#> [271]  0.42245598  0.41341581  0.39922820  0.40536074  0.42275269  0.33978912
#> [277]  0.39810960  0.37110992  0.36267671  0.36708540  0.36927700  0.35926001
#> [283]  0.32982783  0.27392635  0.28180281  0.33069333  0.31153255  0.73464798
#> [289]  0.20596348  0.12541653  0.29920536  0.21105641  0.12893475  0.15280339
#> [295]  0.24339488  0.33317425  0.38677974  0.36324240  0.29328039  0.22440322
#> [301]  0.17596832  0.17802589