Wavelet EbayesThresh Arguments

The main goal of this function is to generate a tibble with combinations of some possible arguments that can be passed to get wavelet-based signal estimation from the function
EbayesThresh::ebayesthresh.wavelet.

wav_args_ebthr(xform, wtlist, ebthrlist)

Arguments

xform
A named list of wavelet transform methods. Types can be "dwt" and "modwt".

wtlist
A named list of arguments to use in the wavelet decomposition, excluding the input vector x, performed by the functions waveslim::dwt or waveslim::modwt.

ebthrlist
A named list of arguments that can be passed to the function
EbayesThresh::ebayesthresh.wavelet excluding xtr.

Value

S3 objetc of class args_ebthr. This object is a list of the following elements:

args_tbl

A tibble with arguments to get wavevel-based signal estimation from
EbayesThresh::ebayesthresh.wavelet

.
nwt_args

A integer vector of length two. The first element is the number of arguments to pass to the function waveslim::dwt or waveslim::modwt. The second element is the number of arguments to the EbayesThresh::ebayesthresh.wavelet

.

See also

Examples

# Wavelet transform

ebthr_xform <- list(wt = c("dwt", "modwt"))

# Wavelet filter (package waveslim)

ebthr_wt <- list(
  wf = c("haar", "d4", "d6", "la8"),
  n.levels = 1:4,
  boundary = c("periodic", "reflection")
)

# Some arguments that can be passed to EbayesThresh::ebayesthresh.wavelet

ebthr_ebwav <- list(
  vscale = c("level", "independet"),
  a = 0.5,
  prior = c("laplace", "cauchy"),
  threshrule = c("median", "mean", "soft", "hard")
)

ebthr_args <- wav_args_ebthr(ebthr_xform, ebthr_wt, ebthr_ebwav)