Wavelet Smooth

This functional is a generic function that computes wavelet-based signals by mapping diferent arguments that can be used directly or indirectly in functions wmta::wavShrink,
EbayesThresh::ebayesthresh.wavelet and wavethresh::threshold. Each estimated signal is a wavelet core inflation measure. wav_smooth calls smooth_wavelet for the estimation of the wavelet-based signal.

wav_smooth(x, y)

Arguments

x
A vector or a time series containing the data (headline inflation).

y
An object of class args_wshr, args_ebthr or args_wthr. Those objects can be obtained through functions wav_args_ebthr, wav_args_wshr and wav_args_wthr.

Value

A tibble containing the estimated wavelet-based signal (x_sm column). Arguments that are not presented in the tibble are the default ones of the respective function that generated the signal (wavelet core inflation).

See also

Examples

wshr_obj <- wav_args_wshr(list(
  wavelet = c("haar", "d4", "d6", "d8", "s8"),
  n.level = 1:4
))

inf_head <- coreinf_br[["ipca"]]

wav_smooth(inf_head, wshr_obj)

#> # A tibble: 20 x 3
#>    wavelet n.level x_sm       
#>    <chr>     <int> <list>     
#>  1 haar          1 <dbl [302]>
#>  2 d4            1 <dbl [302]>
#>  3 d6            1 <dbl [302]>
#>  4 d8            1 <dbl [302]>
#>  5 s8            1 <dbl [302]>
#>  6 haar          2 <dbl [302]>
#>  7 d4            2 <dbl [302]>
#>  8 d6            2 <dbl [302]>
#>  9 d8            2 <dbl [302]>
#> 10 s8            2 <dbl [302]>
#> 11 haar          3 <dbl [302]>
#> 12 d4            3 <dbl [302]>
#> 13 d6            3 <dbl [302]>
#> 14 d8            3 <dbl [302]>
#> 15 s8            3 <dbl [302]>
#> 16 haar          4 <dbl [302]>
#> 17 d4            4 <dbl [302]>
#> 18 d6            4 <dbl [302]>
#> 19 d8            4 <dbl [302]>
#> 20 s8            4 <dbl [302]>