wavcoreinf: estimation and evaluation of wavelet core infaltion measures

wavcoreinf provides a flexible estimation and evaluation of wavelet core inflation from wavelet-based signal estimation methods. Functionals are developed to facilitate the estimation of several specifications permitting the selection of the best one according to specific criteria established in the core inflation literature.

Wavelet-based signal estimation

The wavelet methods for signal estimation are those available in three R packages:

wmtsa

EbayesThresh

wavethresh

The specific methods are explained on the help page of the functions: wmtsa::wavShrink,
EbayesThresh::ebayesthresh.wavelet and wavethresh::threshold.wd

The textbooks by Percival and Walden (2000) and Nason (2008) explain in more details these methods.

Core inflation

Certain criteria that a good core inflation measure should satisfy and which are analyzed in this package include the following:

Standard desviaton

Bias

Trend representation

Dynamic adjustment

Forecasting

Examples using wavelet methods to construct core inflation measure can be viewed in the works of Baqaee (2010) and Dowd and Loh (2011). A general discussion of core inflation measures for the Brazilian case is available in Silva Filho and Figueiredo (2011).

References

Baqaee, David (2010). Using wavelets to measure core inflation: The case of New Zealand, The North American Journal of Economics and Finance, Volume 21, Issue 3, Pages 241-255.

Dowd, K., Cotter, J., & Loh, L. (2011). U.S. core inflation: a wavelet analysis. Macroeconomic Dynamics, 15(4), 513-536.

Nason, G.P. (2008). Wavelet Methods in Statistics with R (1 ed.). Springer Publishing Company, Incorporated.

Percival, D., & Walden, A. (2000). Wavelet Methods for Time Series Analysis (Cambridge Series in Statistical and Probabilistic Mathematics). Cambridge: Cambridge University Press.

Silva Filho, Tito N. Teixeira da, & Figueiredo, Francisco Marcos R. (2011). Has core inflation been doing a good job in Brazil?. Revista Brasileira de Economia, 65(2), 207-233.