Summarize Statistics of the Models by Group

This function calculates some statistics about wavelet core inflation measure by group.

wcore_best_est(x, y, group)

Arguments

x
An objetct generated by the functions wcore_summary, wcore_summary_fcast or wcore_summary_best.

y
A numeric vector (headline inflation) that must be the same x used in wav_smooth.

group
A groupping variable representing a parameter used to compute wavelet-based signal estimation that must be presenting in the object x.

Value

A tibble. The fist column is the groupping variable. The other columns are:

mean2

Mean of the wavelet core inflation.

max

Maximun value of the wavelet core inflation.

min

Maximun value of the wavelet core inflation.

sd2

Standard desviation.

cv

Coeficient of variation of the wavelet core inflation.

tmean

t-tests involving wavelet core inflation and headline inflation.

Examples

library(lubridate)

#> Warning: package 'lubridate' was built under R version 3.6.3

#> 
#> Attaching package: 'lubridate'

#> The following objects are masked from 'package:base':
#> 
#>     date, intersect, setdiff, union

wshr_obj <- wav_args_wshr(list(
wavelet = c("haar", "d4", "d6", "d8", "s8"),
n.level = 1:4
))

inf_head <- coreinf_br[["ipca"]]

date_start <- coreinf_br[["date"]][1]
ts_start <- c(year(date_start), month(date_start))
inf_head_ts <- ts(inf_head, start = ts_start, frequency = 12)

core_wavelet <- wav_smooth(inf_head, wshr_obj)
core_wavelet2 <- wcore_table(core_wavelet, inf_head_ts)

# A more realistic estimation of the forecasting
# error could have h and k greather than 2 and 6.

err_wcore <- error_wave_summary(h = 2, x = wshr_obj,
                                y = inf_head, lags = lags(2,1),
                                k = 6, RMSE = TRUE)

wavcore_smry <- wcore_summary_fcast(inf_head, core_wavelet2, err_wcore)

#> Warning: unnest() has a new interface. See ?unnest for details.
#> Try `df %>% unnest(c(sd, mav25, no_bias, dyn_inf, dyn_core))`, with `mutate()` if needed

x <- wcore_summary_best(wavcore_smry)

wcore_best_est(wavcore_smry, inf_head, wavelet)

#> Warning: unnest() has a new interface. See ?unnest for details.
#> Try `df %>% unnest(c(date, x_sm))`, with `mutate()` if needed

#> # A tibble: 5 x 7
#>   wavelet mean2   max    min   sd2    cv tmean
#>   <chr>   <dbl> <dbl>  <dbl> <dbl> <dbl> <dbl>
#> 1 d4      0.602  6.86 -0.362 0.583 0.968 1.00 
#> 2 d6      0.602  6.81 -0.367 0.582 0.966 0.997
#> 3 d8      0.602  6.84 -0.369 0.581 0.965 1.00 
#> 4 haar    0.603  6.84 -0.340 0.572 0.950 0.996
#> 5 s8      0.602  6.83 -0.369 0.582 0.966 0.999

wcore_best_est(x, inf_head, n.level)

#> Warning: unnest() has a new interface. See ?unnest for details.
#> Try `df %>% unnest(c(date, x_sm))`, with `mutate()` if needed

#> # A tibble: 1 x 7
#>   n.level mean2   max     min   sd2    cv tmean
#>     <int> <dbl> <dbl>   <dbl> <dbl> <dbl> <dbl>
#> 1       4 0.603  6.80 -0.0746 0.559 0.927 0.993