Compute Best (Median) Wavelet Series

wcore_best_median extracts the series of the best model computed by the function wcore_summary_best. If more than one model is selected, this function calculates the median of the models.

wcore_best_median(x)

Arguments

x
An object from wcore_summary_best.

Value

A tibble.

See also

Examples

library(lubridate)

wshr_obj <- wav_args_wshr(list(
wavelet = c("haar", "d4", "d6", "d8", "s8"),
n.level = 1:4
))

inf_head <- coreinf_br[["ipca"]]

date_start <- coreinf_br[["date"]][1]
ts_start <- c(year(date_start), month(date_start))
inf_head_ts <- ts(inf_head, start = ts_start, frequency = 12)

core_wavelet <- wav_smooth(inf_head, wshr_obj)
core_wavelet2 <- wcore_table(core_wavelet, inf_head_ts)

# A more realistic estimation of the forecasting
# error could have h and k greather than 2 and 6.

err_wcore <- error_wave_summary(h = 2, x = wshr_obj,
                                y = inf_head, lags = lags(2,1),
                                k = 6, RMSE = TRUE)

wavcore_smry <- wcore_summary_fcast(inf_head, core_wavelet2, err_wcore)

#> Warning: unnest() has a new interface. See ?unnest for details.
#> Try `df %>% unnest(c(sd, mav25, no_bias, dyn_inf, dyn_core))`, with `mutate()` if needed

x <- wcore_summary_best(wavcore_smry)

wcore_best_median(x)

#> Warning: unnest() has a new interface. See ?unnest for details.
#> Try `df %>% unnest(c(date, x_sm))`, with `mutate()` if needed

#> # A tibble: 302 x 2
#>     date wcore_med
#>    <dbl>     <dbl>
#>  1 1994.      6.80
#>  2 1995.      1.74
#>  3 1995.      1.61
#>  4 1995.      2.70
#>  5 1995.      2.76
#>  6 1995.      1.67
#>  7 1995       1.61
#>  8 1995.      1.50
#>  9 1995.      1.67
#> 10 1995.      2.22
#> # ... with 292 more rows