Includes Forecasting in the Wavelet Core Inflation Summary

wcore_summary_fcast includes forecasting error in the summary of the wavelet core inflation measure. This is done separetely because forecasting can be too consuming in terms of computation time.

wcore_summary_fcast(y, wcoretbl, error)

Arguments

y
A numeric vector (headline inflation) that must be the same x used in wav_smooth.

wcoretbl
An object generated by the function wcore_table.

error
An object generated by the function error_wave_summary.

Value

A tibble. New columns are added to the object wcoretbl:

error

Mean of RMSE for several horizons. See error_wave_summary.

sd, mav25, no_bias, alpha, val_p, alpha1, val_p1

These columns are the same ones explained in the section Value of the wcore_summary function.

Examples

library(lubridate)

wshr_obj <- wav_args_wshr(list(
wavelet = c("haar", "d4", "d6", "d8", "s8"),
n.level = 1:4
))

inf_head <- coreinf_br[["ipca"]]

date_start <- coreinf_br[["date"]][1]
ts_start <- c(year(date_start), month(date_start))
inf_head_ts <- ts(inf_head, start = ts_start, frequency = 12)

core_wavelet <- wav_smooth(inf_head, wshr_obj)
core_wavelet2 <- wcore_table(core_wavelet, inf_head_ts)

# A more realistic estimation of the forecasting
# error could have h and k greather than 2 and 6.

err_wcore <- error_wave_summary(h = 2, x = wshr_obj,
                                y = inf_head, lags = lags(2,1),
                                k = 6, RMSE = TRUE)

wcore_summary_fcast(inf_head, core_wavelet2, err_wcore)

#> Warning: unnest() has a new interface. See ?unnest for details.
#> Try `df %>% unnest(c(sd, mav25, no_bias, dyn_inf, dyn_core))`, with `mutate()` if needed

#> # A tibble: 20 x 12
#>    wavelet n.level x_sm  date  error    sd mav25 no_bias  alpha    val_p alpha1
#>    <chr>     <int> <lis> <lis> <dbl> <dbl> <dbl>   <dbl>  <dbl>    <dbl>  <dbl>
#>  1 haar          1 <dbl~ <dbl~ 0.236 0.594 0.397  0.0377 -1.98  1.29e-38 0.459 
#>  2 d4            1 <dbl~ <dbl~ 0.233 0.602 0.409  0.216  -2.22  6.77e-42 0.588 
#>  3 d6            1 <dbl~ <dbl~ 0.225 0.602 0.409  0.305  -2.09  3.25e-38 0.400 
#>  4 d8            1 <dbl~ <dbl~ 0.224 0.601 0.408  0.508  -1.97  6.90e-44 0.306 
#>  5 s8            1 <dbl~ <dbl~ 0.223 0.601 0.408  0.381  -2.07  2.39e-46 0.362 
#>  6 haar          2 <dbl~ <dbl~ 0.266 0.576 0.370  0.0870 -1.20  1.79e-30 0.125 
#>  7 d4            2 <dbl~ <dbl~ 0.274 0.587 0.388  0.124  -1.43  9.10e-31 0.231 
#>  8 d6            2 <dbl~ <dbl~ 0.271 0.586 0.385  0.169  -1.45  5.90e-37 0.256 
#>  9 d8            2 <dbl~ <dbl~ 0.246 0.585 0.385  0.387  -1.33  1.08e-36 0.122 
#> 10 s8            2 <dbl~ <dbl~ 0.284 0.587 0.389  0.330  -1.41  1.27e-36 0.157 
#> 11 haar          3 <dbl~ <dbl~ 0.278 0.564 0.353  0.126  -0.879 1.26e-21 0.0621
#> 12 d4            3 <dbl~ <dbl~ 0.274 0.577 0.373  0.0594 -1.15  3.10e-25 0.202 
#> 13 d6            3 <dbl~ <dbl~ 0.263 0.575 0.370  0.0797 -1.18  1.14e-29 0.214 
#> 14 d8            3 <dbl~ <dbl~ 0.244 0.574 0.369  0.168  -1.12  3.18e-30 0.109 
#> 15 s8            3 <dbl~ <dbl~ 0.290 0.574 0.369  0.113  -1.14  2.19e-29 0.153 
#> 16 haar          4 <dbl~ <dbl~ 0.271 0.559 0.348  0.140  -0.792 2.27e-19 0.0457
#> 17 d4            4 <dbl~ <dbl~ 0.276 0.569 0.363  0.0327 -1.03  6.31e-23 0.156 
#> 18 d6            4 <dbl~ <dbl~ 0.264 0.567 0.361  0.0282 -1.10  4.43e-27 0.206 
#> 19 d8            4 <dbl~ <dbl~ 0.254 0.568 0.361  0.0716 -1.05  2.37e-27 0.0890
#> 20 s8            4 <dbl~ <dbl~ 0.290 0.569 0.363  0.0363 -1.08  4.03e-26 0.146 
#> # ... with 1 more variable: val_p1 <dbl>